Split declaration and parsing parts for opening hours
